requirements.txt
Django>=4.0,<5.0
djangorestframework
python-dotenv
dj-database-url
psycopg2-binary
django-cors-headers
whitenoise
gunicorn
drf-nested-routers